service/frontend/code/requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
gunicorn==21.2.0
requests==2.32.5

